[REFACTOR] UserAvatarData dto로 변경 후 패키지 이전
