+++ src/main/java/com/wakeUpTogetUp/togetUp/utils/mapper/EntityDtoMapper.java
@@ -4,7 +4,7 @@ import com.wakeUpTogetUp.togetUp.api.alarm.dto.response.AlarmSimpleRes;
 import com.wakeUpTogetUp.togetUp.api.alarm.dto.response.GetAlarmRes;
 import com.wakeUpTogetUp.togetUp.api.alarm.model.Alarm;
 import com.wakeUpTogetUp.togetUp.api.avatar.model.Avatar;
-import com.wakeUpTogetUp.togetUp.api.avatar.vo.UserAvatarData;
+import com.wakeUpTogetUp.togetUp.api.avatar.dto.response.UserAvatarResponse;
 import com.wakeUpTogetUp.togetUp.api.mission.dto.response.GetMissionLogRes;
 import com.wakeUpTogetUp.togetUp.api.mission.dto.response.GetMissionObjectRes;
 import com.wakeUpTogetUp.togetUp.api.mission.dto.response.GetMissionWithObjectListRes;
@@ -12,10 +12,8 @@ import com.wakeUpTogetUp.togetUp.api.mission.model.Mission;
 import com.wakeUpTogetUp.togetUp.api.mission.model.MissionLog;
 import com.wakeUpTogetUp.togetUp.api.mission.model.MissionObject;
 import com.wakeUpTogetUp.togetUp.api.room.dto.response.RoomDetailRes;
-import com.wakeUpTogetUp.togetUp.api.room.dto.response.RoomInfoRes;
 import com.wakeUpTogetUp.togetUp.api.room.dto.response.RoomRes;
 import com.wakeUpTogetUp.togetUp.api.room.dto.response.RoomUserMissionLogRes;
-import com.wakeUpTogetUp.togetUp.api.room.model.Room;
 import com.wakeUpTogetUp.togetUp.api.room.model.RoomUser;
 import com.wakeUpTogetUp.togetUp.api.users.model.UserAvatar;
 
@@ -105,13 +103,13 @@ public interface EntityDtoMapper {
     // Avatar
     @Mapping(source = "avatar.id", target = "avatarId")
     @Mapping(source = "avatar.theme.value", target = "theme")
-    UserAvatarData toUserAvatarData(Avatar avatar, @Context List<UserAvatar> userAvatarList);
+    UserAvatarResponse toUserAvatarData(Avatar avatar, @Context List<UserAvatar> userAvatarList);
 
-    List<UserAvatarData> toUserAvatarDataList(List<Avatar> avatarList,
+    List<UserAvatarResponse> toUserAvatarDataList(List<Avatar> avatarList,
                                               @Context List<UserAvatar> userAvatarList);
 
     @AfterMapping
-    default void setIsUnlockedStatus(@MappingTarget UserAvatarData target, Avatar avatar,
+    default void setIsUnlockedStatus(@MappingTarget UserAvatarResponse target, Avatar avatar,
                                      @Context List<UserAvatar> userAvatarList) {
         target.setIsUnlocked(userAvatarList.stream()
                 .anyMatch(userAvatar -> userAvatar.getAvatar().getId().equals(avatar.getId())));
